fix: replace `"` with `&quot;` in xml attr values
While testing on pcloudy's demo app, `utils.ExtractFirstUniqueXMLID` threw
an error stating that given XML is invalid. This error occurred because
one of attribute had quotes in the value. We tried to escape the value
using `strconv.Quote` but soon figured out that xml doesn't support
escaping values. To resolve this issue, we now replace all the double quotes in the
value with `&quot;`.